+++ doc/examples/nn_training.py
@@ -1,8 +1,8 @@
 import numpy as np
 
-from lightprop.algorithms.nn_trainer import NNTrainer
 from lightprop.calculations import get_gaussian_distribution
 from lightprop.lightfield import LightField
+from lightprop.optimization.nn import NNTrainer
 from lightprop.propagation.params import PropagationParams
 from lightprop.visualisation import Plotter, PlotTypes
